+++ ikea-zigbee-drivers/CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.2] - 2023-11-25
+
+### Changed
+- All: Change battery percentage reporting interval from 11 hours to 5 hours
+
+### Fixed
+- All: Set default values for null values in `updated()` - reported by @leo_charles
+- All: Ignore attribute reports we don't care about (configured by other driver) - reported by @ymerj
+
 ## [3.4.1] - 2023-11-11
 
 ### Fixed